requirements.txt
streamlit>=1.37.1
openai>=1.12.0
beautifulsoup4>=4.12.3
requests>=2.32.3
python-docx>=1.1.2
reportlab>=4.4.1
serpapi>=0.1.0
python-dotenv>=1.1.0
tiktoken>=0.6.0
